+++ greenery/fsm_test.py
@@ -826,53 +826,6 @@ def test_add_anything_else() -> None:
     assert (fsm1 + fsm2).accepts("ba")
 
 
-def test_anything_else_singleton() -> None:
-    assert AnythingElse.TOKEN is ANYTHING_ELSE
-
-
-def test_anything_else_self() -> None:
-    """ANYTHING_ELSE is consistently equal to itself."""
-
-    # pylint: disable=comparison-with-itself
-    # pylint: disable=unneeded-not
-    assert not ANYTHING_ELSE < ANYTHING_ELSE
-    assert ANYTHING_ELSE <= ANYTHING_ELSE
-    assert not ANYTHING_ELSE != ANYTHING_ELSE
-    assert ANYTHING_ELSE == ANYTHING_ELSE
-    assert ANYTHING_ELSE >= ANYTHING_ELSE
-    assert not ANYTHING_ELSE > ANYTHING_ELSE
-
-
-@pytest.mark.parametrize(
-    argnames="val",
-    argvalues=(
-        float("-inf"),
-        float("nan"),
-        float("inf"),
-        0,
-        "abc",
-        object(),
-        str(ANYTHING_ELSE),
-    ),
-)
-def test_anything_else_sorts_after(val: object) -> None:
-    """ANYTHING_ELSE sorts strictly after anything."""
-
-    assert not ANYTHING_ELSE < val
-    assert not ANYTHING_ELSE <= val
-    assert not ANYTHING_ELSE == val
-    assert ANYTHING_ELSE != val
-    assert ANYTHING_ELSE > val
-    assert ANYTHING_ELSE >= val
-
-    assert val < ANYTHING_ELSE
-    assert val <= ANYTHING_ELSE
-    assert val != ANYTHING_ELSE
-    assert not val == ANYTHING_ELSE
-    assert not val > ANYTHING_ELSE
-    assert not val >= ANYTHING_ELSE
-
-
 def test_anything_else_pickle() -> None:
     # [^z]*
     fsm1 = Fsm(
